Models, fixing unique and foreign keys
Multiple keys were foreign key + unique=true. this can be converted to onetoone so I did that

flight_ticket should not have any unique=true fields. later on we need to implement "if flight_id is unique and customer_id is unique" when we enter data in there

flight destination and flight origin were given related_name to try and fix the issue there and allow multiple countries for each flight